requirements.txt
langchain
transformers
torch
fastapi
uvicorn
faiss-cpu
numpy

